Remove old stuff, put Neu in top level
