--- core/src/main/scala/flawless/fixpoint/HFunctor.scala
@@ -1,7 +1,0 @@
-package flawless.fixpoint
-
-import cats.~>
-
-trait HFunctor[F[_[_], _]] {
-  def hmap[G[_], I[_]](nt: G ~> I): F[G, ?] ~> F[I, ?]
-}
