--- core/src/main/scala/flawless/fixpoint/algebra.scala
@@ -1,7 +1,0 @@
-package flawless.fixpoint
-
-import cats.~>
-
-object algebra {
-  type HAlgebra[F[_[_], _], G[_]] = F[G, ?] ~> G
-}
