+++ examples/src/main/scala/flawless/examples/FlakySuite.scala
@@ -4,8 +4,9 @@ import cats.effect.IO
 
 import scala.util.Random
 import cats.implicits._
-import flawless.data.neu._
-import flawless.data.neu.dsl._
+import flawless.dsl._
+import flawless.SuiteClass
+import flawless.data.Suite
 
 object FlakySuite extends SuiteClass[IO] {
   private val flaky = IO(Random.nextInt(10000)).map(_ =!= 0)
